Fixes blocking of query requests when player with name "null" joins
When a player with the name "null" is joining the server then this will cause a java.lang.NullPointerException.
After it the query server fails for new queries and block incoming requests for 300 seconds. In this time the server seems dead/offline, but is currently running.
With this fix those players are kicked at login. Maybe you can solve it in a more elegand way. ;-)